// src/app/types/types.ts
export interface User {
  id: string; 
  userName: string; 
  email: string;
  password: string;
  createdAt?: Date|null
  updatedAt?: Date |null
  news: News[]
  role: string
}
export interface UpdateUser {
  id: string; 
  userName: string; 
  email: string;
  password: string;
  role: string

}
// ...
  export interface News {
    id: string;
    headline: string;
    title: string;
    img: string;
    user: User;
    userId: string;
    department: Department;
    depID: string;
  }
  
  export interface Session {
    id: string;
    expirationDate: Date;
    createdAt: Date;
    updatedAt: Date;
    user: User;
    userId: string;
  }
  
  export interface Department {
    id: string;
    depName: string;
    createdAt: Date;
    updatedAt: Date;
    isDeleted: boolean;
    staffs: Staff[];
    doctors: Doctor[];
    news: News[];
  }